% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tam_par_MCM.R
\name{cal_max_curv_mod}
\alias{cal_max_curv_mod}
\title{Tamanho ótimo de parcela pelo método da Máxima Curvatura Modificada (MCM)}
\usage{
cal_max_curv_mod(Xo, CV, main = "Meier & Lessmann (1971)", plot = TRUE)
}
\arguments{
\item{Xo}{valor numérico correspondente ao tamanho ótimo de parcela de
referência (X₀), obtidos pelo agrupamento de unidades experimentais basicas, 
obtidas em ensaios de uniformidade.}

\item{CV}{coeficiente de variação das parcelas agrupadas, expresso em porcentagem (%).}

\item{main}{título do gráfico gerado pela função.}

\item{plot}{lógico; se \code{TRUE}, gera o gráfico da relação entre o
tamanho da parcela e o coeficiente de variação estimado.}
}
\value{
Um valor numérico correspondente ao tamanho ótimo de parcela
  estimado pelo método da Máxima Curvatura Modificada.
}
\description{
Estima o tamanho ótimo de parcela utilizando o método da Máxima Curvatura
Modificada, proposto por Meier & Lessmann (1971), a partir do tamanho ótimo
de parcela de referência (X0) e do coeficiente de variação do ensaio.
}
\details{
O método baseia-se na identificação do ponto em que o aumento do tamanho
da parcela passa a resultar em ganhos irrisórios na precisão experimental,
sendo amplamente utilizado em estudos de dimensionamento experimental.
}
\examples{
# ---------------------------------------
# Conjunto de dados de exemplo
# ---------------------------------------
dados_curvatura <- data.frame(
  Xo = c(1, 2, 4, 8,
         2, 4, 8, 16,
         3, 6, 12, 24,
         6, 12, 24),
  CV = c(19.55309038, 15.23512639, 10.0533745,  8.203086387,
         14.021032,   12.02564594, 7.707005924, 7.207523603,
         11.51373415, 9.876254603, 11.59399511, 6.690432438,
         7.504701872, 5.791576016, 3.084453732)
)

# Valores médios utilizados no método
Xo_med <- mean(dados_curvatura$Xo)
CV_med <- mean(dados_curvatura$CV)

# Estimativa do tamanho ótimo de parcela
cal_max_curv_mod(
  Xo = Xo_med,
  CV = CV_med,
  plot = TRUE
)
}
\references{
MEIER, V. D.; LESSMAN, K. J. (1971).Estimation of optimum field plot shape and size 
for testing yield in Crambe abyssinica hochst. Crop Science, v. 11, p. 648-650.
}
