% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata.R
\docType{data}
\name{dados_ensaio_C1}
\alias{dados_ensaio_C1}
\title{Dados de um ensaio de uniformidade (Cultivar C1)}
\format{
Um data.frame com 18 observações e 11 variáveis:
\describe{
  \item{Cultivar}{Cultivar avaliado}
  \item{Rep}{Repetição}
  \item{Linha}{Identificação da linha}
  \item{C1--C8}{Variáveis quantitativas avaliadas}
}
}
\source{
Dados simulados para fins didáticos.
}
\usage{
dados_ensaio_C1
}
\description{
Conjunto de dados provenientes de um ensaio de uniformidade,
contendo medições em unidades básicas experimentais organizadas
em linhas, colunas e repetições, utilizado como exemplo
para funções de planejamento experimental.
}
\keyword{datasets}
