% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{DimExp-package}
\alias{DimExp}
\alias{DimExp-package}
\title{DimExp: Tools for Experimental Design Sizing}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions to estimate optimal plot size for experimental design, based on the relationship between the coefficient of variation and plot size. Implements the linear response plateau (LRP) and quadratic response plateau (QRP) models, with standardized diagnostic statistics and plotting.
}
\author{
\strong{Maintainer}: Willyan Bandeira \email{bandeira.wjab@gmail.com} (\href{https://orcid.org/0000-0002-9430-3664}{ORCID})

Authors:
\itemize{
  \item Willyan Bandeira \email{bandeira.wjab@gmail.com} (\href{https://orcid.org/0000-0002-9430-3664}{ORCID})
  \item Leonardo Pradebon \email{leonardopradebon@gmail.com} (\href{https://orcid.org/0000-0001-7827-6312}{ORCID})
  \item Ivan Carvalho \email{carvalho.irc@gmail.com} (\href{https://orcid.org/0000-0001-7947-4900}{ORCID})
  \item Murilo Loro \email{muriloloro@gmail.com} (\href{https://orcid.org/0000-0003-0241-4226}{ORCID})
}

}
\keyword{internal}
