% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrp.R
\name{LRP}
\alias{LRP}
\title{Tamanho ótimo de parcela pelo modelo linear de resposta com platô (LRP)}
\usage{
LRP(data, x_col = "x", cv_col = "cv", plot = TRUE)
}
\arguments{
\item{data}{\code{data.frame} contendo as variáveis de tamanho da parcela
e coeficiente de variação.}

\item{x_col}{Nome da coluna do \code{data.frame} que representa o tamanho
da parcela (padrão: \code{"x"}).}

\item{cv_col}{Nome da coluna do \code{data.frame} que representa o coeficiente
de variação (padrão: \code{"cv"}).}

\item{plot}{se \code{TRUE}, gera o gráfico do modelo ajustado.}
}
\value{
Uma lista contendo:
\describe{
  \item{modelo}{Objeto do ajuste não linear (\code{nls}).}
  \item{parametros}{\code{data.frame} com os parâmetros estimados do modelo,
  incluindo o ponto ótimo de parcela (Xo) e o coeficiente de determinação (R²).}
  \item{predicoes}{\code{data.frame} com valores observados e preditos de CV.}
}
}
\description{
Estima o tamanho ótimo de parcela utilizando o modelo linear de resposta
com platô, proposto por Paranaíba, Ferreira e Morais (2009), a partir do
agrupamento de parcelas de diferentes tamanhos e formas e de seus
respectivos coeficientes de variação (CV).
}
\details{
O modelo é composto por dois segmentos: o primeiro descreve uma relação
linear crescente ou decrescente entre o coeficiente de variação e o
tamanho da parcela até um determinado ponto de quebra (Xo). A partir
desse ponto, observa-se um platô, no qual o coeficiente de variação
assume um valor constante.
}
\examples{
dados <- data.frame(
  x = c(1,2,4,8,2,4,8,16,3,6,12,24,6,12,24),
  cv = c(18.4442670112, 13.7417134862, 10.3923088782, 9.4085177252,
  13.4948645984, 10.4858272400, 6.6816059018, 6.6972274507,
  10.0527295678, 8.0042609703, 11.9884481410, 5.3277132704,
  6.1318167805, 3.8125811769, 0.4788116057))

  a <- LRP(dados)

  #visualizar os resultados
  a
  a$modelo

}
\references{
Paranaíba, P.F.; Ferreira, D. F.; Morais, A. R. (2009) Tamanho ótimo de parcelas
experimentais: proposição de métodos de estimaçãoo. Revista Brasileira de Biometria 27:255–268.
}
