% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lrp.R
\name{plot.lrp_multi}
\alias{plot.lrp_multi}
\title{Plot LRP fits for several trials}
\usage{
\method{plot}{lrp_multi}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"lrp_multi"}.}

\item{...}{further arguments passed to [plot.lrp_fit()].}
}
\value{
A named list of \code{ggplot} objects, invisibly.
}
\description{
Builds one article-style plot per trial. Combine them with a package such as
\pkg{patchwork} or \pkg{ggpubr} for a multi-panel figure.
}
