% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lrp.R
\name{predict.lrp_fit}
\alias{predict.lrp_fit}
\title{Predictions from an LRP fit}
\usage{
\method{predict}{lrp_fit}(object, newx = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{"lrp_fit"}.}

\item{newx}{numeric vector of predictor values. Defaults to the fitted data.}

\item{...}{ignored.}
}
\value{
A numeric vector of predicted responses.
}
\description{
Predictions from an LRP fit
}
