% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mcm.R
\name{predict.mcm_fit}
\alias{predict.mcm_fit}
\title{Predictions from an MCM fit}
\usage{
\method{predict}{mcm_fit}(object, newx = NULL, ...)
}
\arguments{
\item{object}{an \code{"mcm_fit"} object.}

\item{newx}{numeric predictor values; defaults to the fitted data.}

\item{...}{ignored.}
}
\value{
numeric vector of predicted CV values.
}
\description{
Predictions from an MCM fit
}
