% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lrp.R
\name{print.lrp_fit}
\alias{print.lrp_fit}
\title{Print an LRP fit}
\usage{
\method{print}{lrp_fit}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"lrp_fit"}.}

\item{...}{ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an LRP fit
}
