% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lrp.R
\name{print.lrp_multi}
\alias{print.lrp_multi}
\title{Print LRP fits for several trials}
\usage{
\method{print}{lrp_multi}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"lrp_multi"}.}

\item{...}{ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print LRP fits for several trials
}
