% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_trial.R
\name{print.trial_check}
\alias{print.trial_check}
\title{Print a trial check}
\usage{
\method{print}{trial_check}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"trial_check"}.}

\item{...}{ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a trial check
}
