% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lrp.R
\name{summary.lrp_fit}
\alias{summary.lrp_fit}
\title{Summarize an LRP fit}
\usage{
\method{summary}{lrp_fit}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"lrp_fit"}.}

\item{...}{ignored.}
}
\value{
\code{object}, invisibly.
}
\description{
Summarize an LRP fit
}
