% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lrp.R
\name{summary.lrp_multi}
\alias{summary.lrp_multi}
\title{Summarize LRP fits for several trials}
\usage{
\method{summary}{lrp_multi}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"lrp_multi"}.}

\item{...}{ignored.}
}
\value{
\code{object}, invisibly.
}
\description{
Summarize LRP fits for several trials
}
