% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_trial.R
\name{summary.trial_check}
\alias{summary.trial_check}
\title{Summarize a trial check}
\usage{
\method{summary}{trial_check}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"trial_check"}.}

\item{...}{ignored.}
}
\value{
The summary data frame, invisibly.
}
\description{
Summarize a trial check
}
