% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TamParcela.R
\name{calc_paranaiba}
\alias{calc_paranaiba}
\title{Tamanho de parcela}
\usage{
calc_paranaiba(dados, nlin, ncol, nrep, col_inicio = 4, digits = 2)
}
\arguments{
\item{dados}{data.frame ou tibble contendo os dados brutos do ensaio de uniformidade.
Cada linha deve representar uma unidade básica experimental (UEB).
As colunas iniciais devem conter informações de identificação espacial
(por exemplo, linha, coluna e repetição), seguidas pelas variáveis
quantitativas de interesse.}

\item{nlin}{inteiro. Número de linhas do ensaio (direção longitudinal),
correspondente ao número de unidades básicas no sentido das linhas de cultivo.}

\item{ncol}{inteiro. Número de colunas do ensaio (direção transversal),
correspondente ao número de unidades básicas no sentido da largura do experimento.}

\item{nrep}{inteiro. Número de repetições simuladas ou divisões do ensaio,
utilizado para o cálculo das estatísticas de variabilidade entre parcelas.}

\item{col_inicio}{inteiro, opcional. Índica a coluna em que se iniciam
as variáveis quantitativas de interesse em \code{dados}.
As colunas anteriores são tratadas como metadados.
O valor padrão é \code{4}.}

\item{digits}{inteiro, opcional. Número de casas decimais utilizadas
para o arredondamento dos resultados apresentados.
O valor padrão é \code{2}.}
}
\value{
Um data.frame contendo as estatísticas do ensaio de uniformidade,
incluindo médias, coeficientes de variação e parâmetros associados
ao método de Paranaíba para diferentes tamanhos de parcela.
}
\description{
Dimensiona o tamanho ótiomo de parcela para experimentos agrícolas pelo
método da curvatura máxima em função do coeficiente de variação
(PARANAÍBA; FERREIRA; MORAIS, 2009)
}
\details{
O método de Paranaíba baseia-se na relação entre o coeficiente de variação
e o tamanho da parcela, permitindo a estimativa do tamanho ótimo de parcela
e da variabilidade experimental associada. A função utiliza a estrutura
espacial definida por \code{nlin} e \code{ncol} para simular agregações
de unidades básicas experimentais.
}
\examples{
# Carregar os dados de exemplo do pacote
data(dados_ensaio_C1)

# Aplicar o método de Paranaíba
resultado_C1 <- calc_paranaiba(
  dados      = dados_ensaio_C1,
  nlin       = 6,
  ncol       = 8,
  nrep       = 3,
  col_inicio = 4,
  digits     = 3
)

# Visualizar os resultados
resultado_C1

}
\references{
Paranaíba, P F; Ferreira, D F; Morais, A R. (2009). Tamanho ótimo de parcelas
experimentais: proposição de métodos de estimação. \emph{Revista Brasileira
de Biometria}, 27:255–268.
}
